+++ NEWS.md
@@ -2,6 +2,7 @@
 ## v.8.7 - 2025-07-07
 - AWS Athena backend bug fixes 
 - add `temp` option to `@create_table`, default is `true`
+- adjust `case_when` syntax to match TidierData
 
 ## v.8.6 - 2025-05-05
 - add `@pivot_longer`